`CHORE` Code refactoring and minor fixes
